Upload TrainManager.hpp & Add is_newfile to FileManager.hpp

+++ src/TypesAndHeaders.hpp
@@ -7,7 +7,7 @@
 
 namespace sjtu
 {
-	typedef unsigned long long locType;
+	typedef long locType;
 };
 
 #endif //TICKETSYSTEM_TYPESANDHEADERS_HPP
